Move common utilities to separate package to avoid name conflict
